(GH-373) Convert pdkNewTask command to feature
This commit converts pdkNewTaskCommand to be part of
PdkCommandFeature.
