--- src/commands/pdk/pdkNewTaskCommand.ts
@@ -1,35 +1,0 @@
-'use strict';
-
-import * as vscode from 'vscode';
-import { ILogger } from '../../logging';
-import { reporter } from '../../telemetry/telemetry';
-import * as messages from '../../messages';
-
-export class PDKNewTaskCommand {
-  private logger: ILogger;
-  private terminal: vscode.Terminal;
-
-  constructor(logger: ILogger, terminal: vscode.Terminal) {
-    this.logger = logger;
-    this.terminal = terminal;
-  }
-
-  public run() {
-    let nameOpts: vscode.QuickPickOptions = {
-      placeHolder: "Enter a name for the new Puppet Task",
-      matchOnDescription: true,
-      matchOnDetail: true
-    };
-    vscode.window.showInputBox(nameOpts).then(taskName => {
-      this.terminal.sendText(`pdk new task ${taskName}`);
-      this.terminal.show();
-      if (reporter) {
-        reporter.sendTelemetryEvent(messages.PDKCommandStrings.PdkNewTaskCommandId);
-      }
-    })
-  }
-
-  public dispose(): any {
-    this.terminal.dispose();
-  }
-}
